+++ tests/unit/test_models/test_full_battery_models/test_base_battery_model.py
@@ -36,6 +36,7 @@ PRINT_OPTIONS_OUTPUT = """\
 'total interfacial current density as a state': 'false' (possible: ['true', 'false'])
 'working electrode': 'both' (possible: ['both', 'negative', 'positive'])
 'SEI film resistance': 'none' (possible: ['none', 'distributed', 'average'])
+'stress-induced diffusion': 'false' (possible: ['true', 'false'])
 """  # noqa: E501
 
 
@@ -124,8 +125,8 @@ class TestBaseBatteryModel(unittest.TestCase):
             var.x_n: 20,
             var.x_s: 20,
             var.x_p: 20,
-            var.r_n: 30,
-            var.r_p: 30,
+            var.r_n: 20,
+            var.r_p: 20,
             var.y: 10,
             var.z: 10,
             var.R_n: 30,
@@ -185,7 +186,7 @@ class TestBaseBatteryModel(unittest.TestCase):
             pybamm.BaseBatteryModel({"convection": "full transverse"})
         with self.assertRaisesRegex(pybamm.OptionError, "particle"):
             pybamm.BaseBatteryModel({"particle": "bad particle"})
-        with self.assertRaisesRegex(NotImplementedError, "The 'fast diffusion'"):
+        with self.assertRaisesRegex(pybamm.OptionError, "The 'fast diffusion'"):
             pybamm.BaseBatteryModel({"particle": "fast diffusion"})
         with self.assertRaisesRegex(pybamm.OptionError, "particle shape"):
             pybamm.BaseBatteryModel({"particle shape": "bad particle shape"})
@@ -259,6 +260,11 @@ class TestBaseBatteryModel(unittest.TestCase):
                     "plating porosity change"
                 }
             )
+        # stress-induced diffusion
+        with self.assertRaisesRegex(pybamm.OptionError, "cannot have stress"):
+            pybamm.BaseBatteryModel({"stress-induced diffusion": "true"})
+
+        # hydrolysis
         with self.assertRaisesRegex(pybamm.OptionError, "surface formulation"):
             pybamm.lead_acid.LOQS({"hydrolysis": "true", "surface form": "false"})